+++ src/main/java/movieinformation/Form/MovieRegistrationForm.java
@@ -1,13 +1,12 @@
-package MovieInformation.Form;
+package movieinformation.Form;
 
-import MovieInformation.entity.Movie;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.PositiveOrZero;
+import movieinformation.entity.Movie;
+import jakarta.validation.constraints.*;
 
 import java.time.LocalDate;
 
-public class MovieUpdateForm {
+public class MovieRegistrationForm {
+    private int id;
     @NotEmpty //文字列やコレクションなどの文字列が空でないことを検証
     private String name;
     @NotNull //空（null）であるかどうかを検証
@@ -17,7 +16,8 @@ public class MovieUpdateForm {
     @PositiveOrZero //数値が正か 0 であることを検証
     private long boxOffice;
 
-    public MovieUpdateForm(String name, LocalDate releaseDate, String directorName, long boxOffice) {
+    public MovieRegistrationForm(int id, String name, LocalDate releaseDate, String directorName, long boxOffice) {
+        this.id = id;
         this.name = name;
         this.releaseDate = releaseDate;
         this.directorName = directorName;
@@ -25,8 +25,12 @@ public class MovieUpdateForm {
     }
 
     public Movie convertToMovie() {
-        Movie movie = new Movie(this.name, this.releaseDate, this.directorName, this.boxOffice);
+        Movie movie = new Movie(this.id, this.name, this.releaseDate, this.directorName, this.boxOffice);
         return movie;
+    }
+
+    public int getId() {
+        return id;
     }
 
     public String getName() {
