+++ src/main/java/movieinformation/MovieDuplicationException.java
@@ -1,4 +1,4 @@
-package MovieInformation;
+package movieinformation;
 
 public class MovieDuplicationException extends RuntimeException {
     public MovieDuplicationException(String massage) {
