+++ src/main/java/movieinformation/MovieInformationApplication.java
@@ -1,4 +1,4 @@
-package MovieInformation;
+package movieinformation;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
